requirements.txt
flask-cognito-lib
boto3
python-dotenv
gunicorn
pymongo
flask
flask-cors
toml